// src/model/room/Constants.ts
export enum ListType {
    All = "all",
    Today = "today",
    Periodic = "periodic",
    History = "history",
}

export enum RoomType {
    OneToOne = "OneToOne",
    SmallClass = "SmallClass",
    BigClass = "BigClass",
}

export enum RoomStatus {
    Idle = "Idle",
    Started = "Started",
    Paused = "Paused",
    Stopped = "Stopped",
}